requirements.txt
django-filter
django-rest-auth